Fix final CI audit blockers: resolve Rector dry-run conflicts, standardize typed returns, and fix docblock spacing

+++ app/Actions/Workouts/FetchWorkoutsIndexAction.php
@@ -93,7 +93,7 @@ class FetchWorkoutsIndexAction
                     instanceof Carbon
                     ? $date->format('M')
                     : '',
-                'count' => (int) count($rows),
+                'count' => count($rows),
             ])->values();
     }
